Support 128 bit difficulty and 32+48 bit solutions (#14)
* 128 bit difficulty for cpp driver

* Faster uint128_t reverse

* Support 128 bit difficulty for OpenCL

* Regenerate prng if solution wasn't found

* OpenCL fix

* Revert 6869032

* MSVC 128 bit integer support

using Quick C++ Libraries

* Use built-in __int128 if supported

* 64 bit difficulty external support

* Fix Quick C++ Libraries 128 bit bit shift > 64

* 32+ 48 bit solution

* Update gtest

* Use bitwise AND for 48 bit solution part

for better readability

* Merge fix

* Split out 64 bit external difficulty

+++ include/nano_pow/xoroshiro128starstar.hpp
@@ -23,7 +23,7 @@ public:
 		}
 	}
 
-	uint32_t next()
+	uint64_t next()
 	{
 		const uint64_t s0 = s[0];
 		uint64_t s1 = s[1];
